Some changes

+++ Backend/app/business/reservedEvents.manager.ts
@@ -8,12 +8,14 @@ import { NotificationDAO } from "../shared/models/notificationDAO.model";
 import moment from "moment";
 import { NotificationTypeEnum } from "../shared/enums/notificationType.enum";
 import businessContainer from "./business.container";
+import advertisementEventDAO from "../DAO/advertisementEventDAO";
 
 const operations = (context: Context) => {
   const createNewOrUpdate = async (
     reservedData: ReservedAdvertisementEventDAO
   ) => {
     const reserved = await reservedEventDAO.createNewOrUpdate(reservedData);
+    console.log(reserved);
     if (reserved) {
       const dataForTutor: NotificationDAO = {
         userId: reservedData.tutorId,
@@ -22,6 +24,7 @@ const operations = (context: Context) => {
         isReaded: false,
         dateTimeSend: moment().toISOString(),
         typeOfNotification: NotificationTypeEnum.ADVERTISMENTS,
+        advertisementId: reservedData.advertisementId,
       };
 
       const dataForUser: NotificationDAO = {
@@ -31,6 +34,7 @@ const operations = (context: Context) => {
         isReaded: false,
         dateTimeSend: moment().toISOString(),
         typeOfNotification: NotificationTypeEnum.ADVERTISMENTS,
+        advertisementId: reservedData.advertisementId,
       };
 
       const resultForTutor = await businessContainer
@@ -60,8 +64,49 @@ const operations = (context: Context) => {
   };
 
   const removeById = async (advertisementId: string) => {
-    const advertisement = await reservedEventDAO.removeById(advertisementId);
-    if (advertisement) {
+    const reserved = await reservedEventDAO.getReservationById(advertisementId);
+    console.log("reserved", reserved);
+
+    const event = await advertisementEventDAO.findById(
+      reserved.advertisementEventId
+    );
+    console.log(event);
+    console.log("evemt", event);
+    if (event) {
+      console.log("evemt", event);
+
+      const advertisementData = await advertisementDAO.getById(
+        event.advertisementId
+      );
+
+      console.log(advertisementData);
+      const dataForTutor: NotificationDAO = {
+        userId: advertisementData.userId,
+        title: "Rezygnacja z korepetycji",
+        content: "Uzytkownik zrezygnował z wizyty na korepetycje",
+        isReaded: false,
+        dateTimeSend: moment().toISOString(),
+        typeOfNotification: NotificationTypeEnum.ADVERTISMENTS,
+        advertisementId: advertisementId,
+      };
+
+      const dataForUser: NotificationDAO = {
+        userId: reserved.userId,
+        title: "Rezygnacja z korepetycji",
+        content: "Rezygnacja z korepetycji się powiodła",
+        isReaded: false,
+        dateTimeSend: moment().toISOString(),
+        typeOfNotification: NotificationTypeEnum.ADVERTISMENTS,
+        advertisementId: advertisementId,
+      };
+
+      const resultForTutor = await businessContainer
+        .getNotificationManager()
+        .createNewOrUpdate(dataForTutor);
+      const resultForUser = await businessContainer
+        .getNotificationManager()
+        .createNewOrUpdate(dataForUser);
+      const advertisement = await reservedEventDAO.removeById(advertisementId);
       return advertisement;
     }
   };
